[#162] Refactored all component files
* DisplayBox.js and DisplayBox.css moved to DisplayBox directory
* Same for DisplaySensor, Header, Level and Menu
* Boyd Orr, Library and LevelBuilding moved to LevelBuildings directory
* Welcome, Comparator and SSP moved to Pages directory
