+++ frontend/src/components/DisplayBox/DisplayBox.css
