+++ frontend/src/components/DisplayBox/DisplayBox.js
