+++ frontend/src/components/DisplaySensor/DisplaySensor.css
