+++ frontend/src/components/LevelBuildings/BoydOrr.js
@@ -1,16 +1,16 @@
 import React from "react";
 import axios from "axios";
-import Level from "./Level";
-import LevelBuilding from "./LevelBuilding";
-import flash from "../image/flash.png";
+import Level from "../Level/Level.js";
+import LevelBuilding from "./LevelBuilding.js";
+import flash from "../../image/flash.png";
 
-export default class Library extends LevelBuilding {
+export default class BoydOrr extends LevelBuilding {
   componentDidMount() {
     this.setState({
-      url: "library",
-      collection: "Library",
+      url: "boyd-orr",
+      collection: "BoydOrr",
     });
-    axios.get("http://127.0.0.1:5000/library").then((res) => {
+    axios.get("http://127.0.0.1:5000/boyd-orr").then((res) => {
       const all_sensors = res.data;
       this.setState((state) => {
         return { sensors: all_sensors };
@@ -23,7 +23,9 @@ export default class Library extends LevelBuilding {
     const power = this.calculatePowerConsumption();
     return (
       <div>
-        <h2><center>The Library</center></h2>
+        <h2>
+          <center>Boyd Orr Building</center>
+        </h2>
         <center>
           <div style={{backgroundColor:"whitesmoke", width: 300, border: "2px solid" }}>
             <img
